Rename inputdata to parms
